[FIX] ERD 변경에 따른 Cart 도메인 수정
